Work around nested panic in DRM
When a kernel panic occurs while DRM graphics are enabled a second panic
occurs before the message can be printed or a backtrace produced.  This
prevents the users from seeing the panic message.

This change works around the panic in two parts:
 - dont wait for vblank during panic.
 - avoid call to drm_modeset_drop_locks as well due to unknown bug

This should really be done in a FreeBSD-specific layer.

Fixes #1564