tda19988: Add a workaround for the HPD issue
The TDA19988 transmitter in Morello SoC r0p0 and r0p1 might not
correctly detect hot-plugging due to incorrect voltage levels on the HPD
pin.

As a workaround for this issue, a user can set hw.tda19988.broken_hpd to
enforce using a screen if the transmitter only receives an RxSense
packet indicating a screen is attached, ignoring if it's actually able
to receive HDMI input.

Co-authored-by: Jessica Clarke <jrtc27@jrtc27.com>